+++ docs/changelog.md
@@ -6,6 +6,7 @@
 ##  VERSION 1.05
 + Added new outbound message type "constant"
 + Added option to delay encoders (parameter "delay") 
++ Display ACF-Filename in profile window
 
 -----------------------------------------------------------------------------------------------------------------------